Improve cart + reorganize components structure
