+++ Frontend/package-downloader-react/src/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Box, Container, Typography } from '@mui/material';
-import SearchForm from './components/SearchForm';
-import SearchResults from './components/SearchResultsList';
-import PackageCart from './components/PackageCart';
-import DownloadButton from './components/DownloadButton';
+import SearchForm from './components/SearchForm/SearchForm';
+import SearchResults from './components/SearchForm/SearchResultsList';
+import PackageCart from './components/Cart/PackageCart';
+import DownloadPackagesButton from './components/Cart/DownloadButton';
 import { packagesSearchStore } from './stores/PackagesStore';
 
 const App: React.FC = () => {
@@ -16,25 +16,7 @@ const App: React.FC = () => {
       </Typography>
 
       <SearchForm />
-
-      <Box
-        sx={{
-          position: 'fixed',
-          bottom: 0, // Зафиксировать внизу
-          right: 0,  // Зафиксировать вправо
-          width: '300px', // Ширина корзины
-          zIndex: 1000, // Обеспечивает, что корзина выше других элементов
-          backgroundColor: 'white', // Цвет фона, чтобы выделить элемент
-          boxShadow: 3, // Добавить тень для эффекта
-          padding: 2,
-        }}
-      >
-        <Typography variant="h6" gutterBottom>
-          Selected Packages
-        </Typography>
-        <PackageCart />
-        <DownloadButton />
-      </Box>
+      <PackageCart/>
 
       <Typography variant="h6" gutterBottom>
         Search Results
